app/requirements-test.txt
# Requirements for testing Patient Monitoring System

# Core testing framework
pytest>=7.4.0
pytest-django>=4.5.2

# Coverage reporting
pytest-cov>=4.1.0
coverage>=7.2.0

# Database testing
pytest-mock>=3.11.1
factory-boy>=3.3.0

# Performance testing
pytest-benchmark>=4.0.0

# Parallel testing
pytest-xdist>=3.3.1

# HTML reports
pytest-html>=3.1.1

# Additional testing utilities
freezegun>=1.2.2  # For time-based testing
responses>=0.23.1  # For mocking HTTP responses

# Development dependencies
django-extensions>=3.2.0  # For shell_plus and other utilities